Deal with non-printable characters when rendering raw results
